pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["backend*"]

[project]
name = "wd-hydrus-tagger"
version = "0.1.0"
description = "WD14 Tagger integration for Hydrus Network via Web UI"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.10",
    "pyyaml>=6.0.1",
    "httpx>=0.28",
    "onnxruntime>=1.20",
    "Pillow>=10.4",
    "numpy>=1.26",
    "huggingface-hub>=0.26",
    "websockets>=13.0",
]

[project.optional-dependencies]
gpu = ["onnxruntime-gpu>=1.20"]
# Linux: faster asyncio event loop for Hydrus I/O + WebSockets (no effect on Windows/macOS).
perf = ["uvloop>=0.21; sys_platform == 'linux'"]
dev = [
    "httpx>=0.28",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
# Default run (``pytest`` / ``./wd-hydrus-tagger.sh test``): full suite — no ``-m`` filter.
markers = [
    "core: fast unit and HTTP API tests (exclude WebSocket + UI modules for targeted runs)",
    "ws: WebSocket tagging (``tagger_ws``) and recovery helpers",
    "ui: frontend English / static asset checks",
    "slow: heavier scenarios (multi-batch WS, learning calibration, full-tree UI scan)",
    "full: complete suite — every test module includes this; ``pytest -m full`` runs all tests with the default coverage gate",
]
addopts = [
    "--cov=backend",
    "--cov-report=term-missing:skip-covered",
]

[tool.coverage.run]
branch = true
source = ["backend"]
omit = [
    "*/tests/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "\\.\\.\\.",
]
precision = 1
skip_empty = true
# Total line+branch coverage for ``backend`` (``pytest`` with dev extras).
# ~82% total: large remainder is mostly ``tagger_ws`` long-run paths; raise toward 85% with more WS scenarios.
fail_under = 82

[project.scripts]
wd-hydrus-tagger = "backend.app:main"
